gutread-rdf-import/requirements.txt
requests
urllib3
bs4
lxml
psycopg2-binary
dotenv